+++ tests/linen/module_test.py
@@ -253,21 +253,8 @@ class ModuleTest(absltest.TestCase):
         return x + self.bias
     x = jnp.array([1.])
     scope = Scope({}, {'params': rngkey}, mutable=['params'])
-    with self.assertRaisesRegex(ValueError, 'bias already in use'):
-      y = Dummy(x.shape, parent=scope)(x)
-
-  def test_setup_var_collision(self):
-    rngkey = jax.random.PRNGKey(0)
-    class Dummy(nn.Module):
-      xshape: Tuple[int]
-      def setup(self):
-        self.bias = self.param('bias', initializers.ones, self.xshape)
-        self.bias = self.param('bias', initializers.ones, self.xshape)
-      def __call__(self, x):
-        return x + self.bias
-    x = jnp.array([1.])
-    scope = Scope({}, {'params': rngkey}, mutable=['params'])
-    with self.assertRaisesRegex(ValueError, 'bias already in use'):
+    msg = 'Could not create param "bias" in Module Dummy: Name in use'
+    with self.assertRaisesRegex(errors.NameInUseError, msg):
       y = Dummy(x.shape, parent=scope)(x)
 
   def test_call_var_collision(self):
@@ -281,7 +268,23 @@ class ModuleTest(absltest.TestCase):
         return x + bias
     x = jnp.array([1.])
     scope = Scope({}, {'params': rngkey}, mutable=['params'])
-    with self.assertRaisesRegex(ValueError, 'bias already in use'):
+    msg = 'Could not create param "bias" in Module Dummy: Name in use'
+    with self.assertRaisesRegex(errors.NameInUseError, msg):
+      y = Dummy(x.shape, parent=scope)(x)
+
+  def test_setup_var_collision(self):
+    rngkey = jax.random.PRNGKey(0)
+    class Dummy(nn.Module):
+      xshape: Tuple[int]
+      def setup(self):
+        self.bias = self.param('bias', initializers.ones, self.xshape)
+        self.bias = self.param('bias', initializers.ones, self.xshape)
+      def __call__(self, x):
+        return x + self.bias
+    x = jnp.array([1.])
+    scope = Scope({}, {'params': rngkey}, mutable=['params'])
+    msg = 'Could not create param "bias" in Module Dummy: Name in use'
+    with self.assertRaisesRegex(errors.NameInUseError, msg):
       y = Dummy(x.shape, parent=scope)(x)
 
   def test_setattr_name_var_disagreement_allowed_in_lists(self):
@@ -320,43 +323,66 @@ class ModuleTest(absltest.TestCase):
     y = Dummy(x.shape, parent=scope)(x)
     self.assertEqual(y, jnp.array([2.]))
 
-  def test_submodule_var_collision(self):
+  def test_submodule_var_collision_with_scope(self):
     rngkey = jax.random.PRNGKey(0)
+
     class Dummy(nn.Module):
       xshape: Tuple[int]
+
       def setup(self):
         self.bias = self.param('bias', initializers.ones, self.xshape)
         self.bias = DummyModule()
+
       def __call__(self, x):
         return x + self.bias
+
     x = jnp.array([1.])
     scope = Scope({}, {'params': rngkey}, mutable=['params'])
-    msg = r'Duplicate use of scope name: "bias"'
-    with self.assertRaisesRegex(errors.ScopeNameInUseError, msg):
+
+    msg = 'Duplicate use of scope name: "bias"'
+    with self.assertRaisesWithLiteralMatch(ValueError, msg):
       y = Dummy(x.shape, parent=scope)(x)
+
+  def test_submodule_var_collision_with_submodule(self):
+    rngkey = jax.random.PRNGKey(0)
+
     class Dummy(nn.Module):
       xshape: Tuple[int]
+
       def setup(self):
         self.bias = self.param('bias', initializers.ones, self.xshape)
+
       @compact
       def __call__(self, x):
         bias = DummyModule(name='bias')
         return x + self.bias
+
     x = jnp.array([1.])
     scope = Scope({}, {'params': rngkey}, mutable=['params'])
-    with self.assertRaisesRegex(ValueError, 'name bias exists already'):
+
+    msg = 'Could not create submodule "bias" in Module Dummy: Name in use'
+    with self.assertRaisesRegex(errors.NameInUseError, msg):
       y = Dummy(x.shape, parent=scope)(x)
+
+  def test_submodule_var_collision_with_params(self):
+    rngkey = jax.random.PRNGKey(0)
+
     class Dummy(nn.Module):
       xshape: Tuple[int]
+
       def setup(self):
         self.bias = DummyModule()
+
       @compact
       def __call__(self, x):
         bias = self.param('bias', initializers.ones, self.xshape)
         return x + self.bias
+
     x = jnp.array([1.])
     scope = Scope({}, {'params': rngkey}, mutable=['params'])
-    with self.assertRaisesRegex(ValueError, 'bias already'):
+
+    msg = 'Could not create param "bias" in Module Dummy: Name in use'
+    with self.assertRaisesRegex(errors.NameInUseError, msg):
       y = Dummy(x.shape, parent=scope)(x)
 
   def test_attr_param_name_collision(self):
@@ -369,7 +395,8 @@ class ModuleTest(absltest.TestCase):
         return x + self.bias
     x = jnp.array([1.])
     scope = Scope({}, {'params': rngkey}, mutable=['params'])
-    with self.assertRaisesRegex(ValueError, 'Name bias already in use'):
+    msg = 'Could not create param "bias" in Module Dummy: Name in use'
+    with self.assertRaisesRegex(errors.NameInUseError, msg):
       y = Dummy(x.shape, parent=scope)(x)
 
   def test_attr_submodule_name_collision(self):
@@ -382,11 +409,13 @@ class ModuleTest(absltest.TestCase):
         return self.bias(x)
     x = jnp.array([1.])
     scope = Scope({}, {'params': rngkey}, mutable=['params'])
-    with self.assertRaisesRegex(ValueError, 'bias exists already'):
+    msg = 'Could not create submodule "bias" in Module Dummy: Name in use'
+    with self.assertRaisesRegex(errors.NameInUseError, msg):
       y = Dummy(x.shape, parent=scope)(x)
 
   def test_only_one_compact_method(self):
-    with self.assertRaisesRegex(RuntimeError, '@compact'):
+    msg = 'Only one method per class can be @compact'
+    with self.assertRaisesRegex(errors.MultipleMethodsCompactError, msg):
       class Dummy(nn.Module):
         @compact
         def call1(self):
@@ -424,7 +453,9 @@ class ModuleTest(absltest.TestCase):
         x = bar(x)
         x = bar(x)
         return x
-    with self.assertRaisesRegex(ValueError, '@compact'):
+    msg = (r'Submodule Dense must be defined in `setup\(\)` or in a method '
+            'wrapped in `@compact`')
+    with self.assertRaisesRegex(errors.AssignSubModuleError, msg):
       Foo().init(random.PRNGKey(0), jnp.ones((1, 3)))
 
   def test_forgotten_compact_annotation_with_explicit_parent(self):
@@ -440,7 +471,9 @@ class ModuleTest(absltest.TestCase):
         x = bar(x)
         return x
 
-    with self.assertRaisesRegex(ValueError, '@compact'):
+    msg = (r'Submodule Dense must be defined in `setup\(\)` or in a method '
+            'wrapped in `@compact`')
+    with self.assertRaisesRegex(errors.AssignSubModuleError, msg):
       Foo().init(random.PRNGKey(0), jnp.ones((1, 3)))
 
   def test_numpy_array_shape_class_args(self):
@@ -568,7 +601,8 @@ class ModuleTest(absltest.TestCase):
 
   def test_module_with_scope_is_not_hashable(self):
     module_a = nn.Dense(10, parent=Scope({}))
-    with self.assertRaisesWithLiteralMatch(ValueError, 'Can\'t call __hash__ on modules that hold variables.'):
+    msg = 'Can\'t call __hash__ on modules that hold variables.'
+    with self.assertRaisesWithLiteralMatch(ValueError, msg):
       hash(module_a)
 
   def test_module_trace(self):
@@ -615,9 +649,38 @@ class ModuleTest(absltest.TestCase):
     self.assertEqual(trace, expected_trace)
 
 
+  def test_module_apply_method(self):
+    class Foo(nn.Module):
+      @nn.compact
+      def __call__(self):
+        pass
+      
+      def test(self):
+        pass
+
+    # We can use both instance and class methods in apply.
+    Foo().apply({}, method=Foo.test)
+    Foo().apply({}, method=Foo().test)
+
+    # We also use a function that is not in the provided Module, although it
+    # should have a first argument representing an instance of the Module (Foo
+    # in this case).
+    x = Foo().apply({}, method=lambda foo_instance: foo_instance)
+    self.assertEqual(type(x), type(Foo()))
+
+    # This is not allowed.
+    msg = 'Cannot call apply()'
+    with self.assertRaisesRegex(errors.ApplyModuleInvalidMethodError, msg):
+      Foo().apply({}, method=lambda: True)
+
+    with self.assertRaisesRegex(errors.ApplyModuleInvalidMethodError, msg):
+      Foo().apply({}, method='allowed_apply_fn')
+
+
   def test_call_unbound_compact_module_methods(self):
     dense = Dense(3)
-    with self.assertRaisesRegex(ValueError, "compact.*unbound module"):
+    msg = r'Can\'t call compact methods on unbound modules'
+    with self.assertRaisesRegex(errors.CallCompactUnboundModuleError, msg):
       dense(jnp.ones((1, )))
 
 
@@ -660,22 +723,23 @@ class ModuleTest(absltest.TestCase):
 
     empty = EmptyModule()
     # It's fine to call methods of unbound methods that don't depend on
-    # attributes defined during `setup`
+    # attributes defined during `setup`.
     self.assertEqual(empty.bar(), 3)
 
 
-  def test_call_unbound_noncompact_module_methods(self):
+  def test_call_unbound_noncompact_module_methods_depending_on_setup(self):
     class EmptyModule(nn.Module):
-      foo: int = 3
+      def setup(self):
+        self.foo = 2
 
       def bar(self):
         return self.foo
 
     empty = EmptyModule()
-    # It's fine to call methods of unbound methods that don't depend on
-    # attributes defined during `setup`
-    self.assertEqual(empty.bar(), 3)
-
+    msg = r'"EmptyModule" object has no attribute "foo"'
+    with self.assertRaisesRegex(AttributeError, msg):
+      empty.bar()
+     
 
   def test_module_with_attrs(self):
     class Foo(nn.Module):
@@ -700,8 +764,11 @@ class ModuleTest(absltest.TestCase):
       def __call__(self):
         self.i = 2  # This is not allowed.
 
-    with self.assertRaisesWithLiteralMatch(TypeError, "Module instance is frozen outside of setup method."):
+    msg = ('Can\'t set i=2 for Module of type Foo: Module instance is frozen '
+           'outside of setup method.')
+    with self.assertRaisesRegex(errors.SetAttributeFrozenModuleError, msg):
       Foo().init(random.PRNGKey(0))
+
 
   def test_compact_module_frozen(self):
     class Foo(nn.Module):
@@ -709,8 +776,11 @@ class ModuleTest(absltest.TestCase):
       def __call__(self):
         self.i = 2
 
-    with self.assertRaisesWithLiteralMatch(TypeError, "Module instance is frozen outside of setup method."):
+    msg = ('Can\'t set i=2 for Module of type Foo: Module instance is frozen '
+           'outside of setup method.')
+    with self.assertRaisesRegex(errors.SetAttributeFrozenModuleError, msg):
       Foo().init(random.PRNGKey(0))
+
 
   def test_submodule_frozen(self):
     class Foo(nn.Module):
@@ -719,7 +789,9 @@ class ModuleTest(absltest.TestCase):
         dense = nn.Dense(10)
         dense.features = 20  # <--- This is not allowed
 
-    with self.assertRaisesWithLiteralMatch(TypeError, "Module instance is frozen outside of setup method."):
+    msg = ('Can\'t set features=20 for Module of type Dense: Module instance '
+           'is frozen outside of setup method.')
+    with self.assertRaisesRegex(errors.SetAttributeFrozenModuleError, msg):
       Foo().init(random.PRNGKey(0))
 
 
@@ -727,10 +799,11 @@ class ModuleTest(absltest.TestCase):
     class Foo(nn.Module):
       pass
 
-    foo = Foo()
-    with self.assertRaisesWithLiteralMatch(AttributeError, "'Foo' object has no attribute '__call__'"):
-      foo.init(random.PRNGKey(0))
-  
+    msg = '"Foo" object has no attribute "__call__"'
+    with self.assertRaisesRegex(AttributeError, msg):
+      Foo().init(random.PRNGKey(0))
+
+
   def test_is_mutable_collection(self):
     class EmptyModule(nn.Module):
       def __call__(self):
@@ -795,7 +868,8 @@ class ModuleTest(absltest.TestCase):
       def setup(self):
         self.c = nn.Dense(2)
 
-    with self.assertRaisesWithLiteralMatch(AttributeError, "'B' object has no attribute 'c'"):
+    msg = '"B" object has no attribute "c"'
+    with self.assertRaisesRegex(AttributeError, msg):
       A().init(random.PRNGKey(0))
 
   def test_unbound_setup_call(self):